Integrate commitments

+++ src/SFA.DAS.ManageApprenticeships.Api/Models/Transfers/GetFinancialBreakdownResponse.cs
@@ -2,7 +2,7 @@
 {
     public class GetFinancialBreakdownResponse
     {
-        public long Commitments { get; set; }
+        public decimal Commitments { get; set; }
         public long ApprovedPledgeApplications { get; set; }
         public long AcceptedPledgeApplications { get; set; }
         public long PledgeOriginatedCommitments { get; set; }
